CreateResearchBot/requirements.txt
vk-api==11.10.0
openai==2.36.0
qdrant-client==1.18.0
python-dotenv==1.2.2
tiktoken==0.12.0
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
sentence-transformers==3.4.1
groq==0.28.0
